+++ Lua/LuaAdapterDefinitions.h
@@ -444,9 +444,9 @@ namespace RTE {
 		/// Schedule to draw a polygon primitive.
 		/// </summary>
 		/// <param name="primitiveMan">A reference to PrimitiveMan, provided by Lua.</param>
-		/// <param name="centerPos">Position of primitive's center in scene coordinates.</param>
+		/// <param name="centerPos">Position of primitive's center in Scene coordinates.</param>
 		/// <param name="color">Color to draw primitive with.</param>
-		/// <param name="vertices">A vector that contains the positions of the primitive's vertices, relative to the center position.</param>
+		/// <param name="verticesTable">A Lua table that contains the positions of the primitive's vertices, relative to the center position.</param>
 		static void DrawPolygonPrimitive(PrimitiveMan &primitiveMan, const Vector &centerPos, int color, const luabind::object &verticesTable);
 
 		/// <summary>
@@ -454,18 +454,18 @@ namespace RTE {
 		/// </summary>
 		/// <param name="primitiveMan">A reference to PrimitiveMan, provided by Lua.</param>
 		/// <param name="player">Player screen to draw primitive on.</param>
-		/// <param name="centerPos">Position of primitive's center in scene coordinates.</param>
+		/// <param name="centerPos">Position of primitive's center in Scene coordinates.</param>
 		/// <param name="color">Color to draw primitive with.</param>
-		/// <param name="vertices">A vector that contains the positions of the primitive's vertices, relative to the center position.</param>
+		/// <param name="verticesTable">A Lua table that contains the positions of the primitive's vertices, relative to the center position.</param>
 		static void DrawPolygonPrimitiveForPlayer(PrimitiveMan &primitiveMan, int player, const Vector &centerPos, int color, const luabind::object &verticesTable);
 
 		/// <summary>
 		/// Schedule to draw a filled polygon primitive.
 		/// </summary>
 		/// <param name="primitiveMan">A reference to PrimitiveMan, provided by Lua.</param>
-		/// <param name="startPos">Start position of the primitive in scene coordinates.</param>
+		/// <param name="startPos">Start position of the primitive in Scene coordinates.</param>
 		/// <param name="color">Color to draw primitive with.</param>
-		/// <param name="vertices">A vector that contains the positions of the primitive's vertices, relative to the center position.</param>
+		/// <param name="verticesTable">A Lua table that contains the positions of the primitive's vertices, relative to the center position.</param>
 		static void DrawPolygonFillPrimitive(PrimitiveMan &primitiveMan, const Vector &startPos, int color, const luabind::object &verticesTable);
 
 		/// <summary>
@@ -473,9 +473,9 @@ namespace RTE {
 		/// </summary>
 		/// <param name="primitiveMan">A reference to PrimitiveMan, provided by Lua.</param>
 		/// <param name="player">Player screen to draw primitive on.</param>
-		/// <param name="startPos">Start position of the primitive in scene coordinates.</param>
+		/// <param name="startPos">Start position of the primitive in Scene coordinates.</param>
 		/// <param name="color">Color to draw primitive with.</param>
-		/// <param name="vertices">A vector that contains the positions of the primitive's vertices, relative to the center position.</param>
+		/// <param name="verticesTable">A Lua table that contains the positions of the primitive's vertices, relative to the center position.</param>
 		static void DrawPolygonFillPrimitiveForPlayer(PrimitiveMan &primitiveMan, int player, const Vector &startPos, int color, const luabind::object &verticesTable);
 
 		/// <summary>
