+++ System/StandardIncludes.h
@@ -77,4 +77,22 @@
 #include <random>
 #include <array>
 #include <filesystem>
+
+namespace std {
+
+	/// <summary>
+	/// Custom std::hash specialization to allow using std::array as key in hash table based containers.
+	/// </summary>
+	template <typename Type, size_t Size> struct hash<array<Type, Size>> {
+		size_t operator()(const array<Type, Size> &arr) const {
+			hash<Type> hasher;
+			size_t outHash = 0;
+			for (size_t i = 0; i < Size; ++i) {
+				// Stolen from java.lang.String.hashCode. It seems to be a popular number, because it is prime, and 31 * x can be implemented quite efficiently as (x << 5) - x.
+				outHash = outHash * 31 + hasher(arr[i]);
+			}
+			return outHash;
+		}
+	};
+}
 #endif
